requirements.txt
Django==1.7a1
MySQL-python==1.2.5
django-tastypie==0.11.0
jsonfield==0.9.20
nose==1.3.0
